requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
streamlit==1.39.0
google-generativeai==0.8.3
python-dotenv==1.0.1
